Further fixes to error when modifying opened files on windows

+++ featherstore/_utils.py
@@ -1,9 +1,10 @@
 import os
-import platform
-import shutil
-import re
-import ctypes
 from pathlib import Path
+import shutil
+import platform
+import subprocess
+import ctypes
+import re
 
 DB_MARKER_NAME = ".featherstore"
 DEFAULT_ARROW_INDEX_NAME = "__index_level_0__"
@@ -27,13 +28,13 @@ def delete_folder_tree(path, db_path):
     if _is_in_database(path, db_path):
         __delete_folder_tree(path)
     else:
-        raise PermissionError("Not a database!")
+        raise PermissionError(f"Can't delete files outside the database ({path})")
 
 
 def _is_in_database(path, db_path):
     path = Path(path)
     db_path = Path(db_path)
-    return db_path in path.parents
+    return db_path in path.parents or path == db_path
 
 
 def __delete_folder_tree(path):
@@ -43,9 +44,13 @@ def __delete_folder_tree(path):
         pass
     except PermissionError as e:
         # Force delete stubborn open file on Windows
-        os.system(f'cmd /k "del /f /q /a {e.filename}"')
-        # Try to delete folder with stubborn file deleted
-        __delete_folder_tree(path)
+        cmd = ["del", "/f", "/a", f"{e.filename}"]
+        output = subprocess.run(cmd, shell=True, check=True, capture_output=True).stderr.decode()
+        if output.startswith('The process cannot access the file'):
+            raise e
+        else:
+            # Try to delete folder with stubborn file deleted
+            __delete_folder_tree(path)
 
 
 def expand_home_dir_modifier(path):
